Added a progress requirement for quests
* This is more or less an optional reporting mechanism for quests with
requirements to be reported

+++ OpenORPG.Database/Models/UserQuestInfo.cs
@@ -23,13 +23,17 @@ namespace Server.Game.Database.Models
         public long QuestId { get; set; }
 
         /// <summary>
-        /// This marks the current state of a quest for a user
+        /// This marks the current state of a quest for a user.
+        /// 
+        /// Since quests that are finished are possible to repeat, this flag could be a few different things.
         /// </summary>
         public QuestState State { get; set; }
 
         /// <summary>
         /// A simple integer used to track the amount of mobs kill in relation to the
-        /// monsters requirement
+        /// monsters requirement.
+        /// 
+        /// This could be eventually replaced with something a bit more generic
         /// </summary>
         [Required, DefaultValue(0)]
         public long MobsKilled { get; set; }
@@ -37,7 +41,7 @@ namespace Server.Game.Database.Models
         /// <summary>
         /// The quest progress ID that the user has moved along to.
         /// </summary>
-        public long QuestProgress { get; set; }
+        public int QuestProgress { get; set; }
 
         [Required]
         public virtual UserHero UserHero { get; set; }
